1.42: Main page cleanup. Taller table. Hide Central Selector Button when toggling the lights table Fixed a bug where the table selection checkmarks were getting ruined when scrolling on and off the table edges.
1.43:
Added Switches to lights table. You can now turn the lights on and off individually without resorting to the official LIFX app.
Remember, the check-mark on the right means the light is selected for action. The switch on the left turns it on and off. If a light is selected (i.e. check-marked), then it will turn itself on when performing an effect, even if it was turned off previously. Checkmark > Switch.

1.41:
Fixed bug where all bulbs were turning on.